tests for queryBuilder
