--- packages/aor-graphql-client-graphcool/src/queryBuilder.test.js
@@ -1,3 +1,0 @@
-test('runs', () => {
-    expect(true).toEqual(true);
-});
